[#1] ShazamService 구현
