+++ AGAMI/.swiftlint.yml
@@ -1,6 +1,7 @@
 disabled_rules:
   - line_length
   - trailing_whitespace
+  - trailing_newline
 
 opt_in_rules:
   - empty_count
